# handle the case smoother when all endpoints are fading in