use weighted shifting instead of shifting to oldest when there are no available stable endpoints