@AlexanderYastrebov 2020-10-20
I did a synthetic test for fresh start case:
```
$ tail -n 10 /etc/hosts
127.0.0.1 h0.test
127.0.0.1 h1.test
127.0.0.1 h2.test
127.0.0.1 h3.test
127.0.0.1 h4.test
127.0.0.1 h5.test
127.0.0.1 h6.test
127.0.0.1 h7.test
127.0.0.1 h8.test
127.0.0.1 h9.test

$ cat ~/tmp/algo.eskip 
r0: Host("h0.test") -> status(200) -> <shunt>;
r1: Host("h1.test") -> status(210) -> <shunt>;
r2: Host("h2.test") -> status(220) -> <shunt>;
r3: Host("h3.test") -> status(230) -> <shunt>;
r4: Host("h4.test") -> status(240) -> <shunt>;
r5: Host("h5.test") -> status(250) -> <shunt>;
r6: Host("h6.test") -> status(260) -> <shunt>;
r7: Host("h7.test") -> status(270) -> <shunt>;
r8: Host("h8.test") -> status(280) -> <shunt>;
r9: Host("h9.test") -> status(290) -> <shunt>;

test1: Path("/test") -> setPath("/") -> fadeIn("3m") -> 
<
    roundRobin, // algorithm
    "http://h0.test:9090",
    "http://h1.test:9090",
    "http://h2.test:9090",
    "http://h3.test:9090",
    "http://h4.test:9090",
    "http://h5.test:9090",
    "http://h6.test:9090",
    "http://h7.test:9090",
    "http://h8.test:9090",
    "http://h9.test:9090"
>;

bin/skipper -routes-file ~/tmp/algo.eskip -access-log-disabled
```

On master branch (just after skipper start) there is:
```
$ echo "GET http://localhost:9090/test" | vegeta attack -duration=20s -rate=100/s | vegeta report;
Requests      [total, rate, throughput]         2000, 100.05, 100.04
Duration      [total, attack, wait]             19.991s, 19.99s, 1.295ms
Latencies     [min, mean, 50, 90, 95, 99, max]  290.588µs, 1.299ms, 1.33ms, 1.591ms, 1.735ms, 2.884ms, 6.153ms
Bytes In      [total, mean]                     0, 0.00
Bytes Out     [total, mean]                     0, 0.00
Success       [ratio]                           100.00%
Status Codes  [code:count]                      200:1859  210:16  220:11  230:15  240:14  250:18  260:17  270:13  280:23  290:14  
Error Set:
```
a clear skew towards first endpoint.

On this branch it looks fine:
```
$ echo "GET http://localhost:9090/test" | vegeta attack -duration=20s -rate=100/s | vegeta report;
Requests      [total, rate, throughput]         2000, 100.05, 100.04
Duration      [total, attack, wait]             19.991s, 19.99s, 1.512ms
Latencies     [min, mean, 50, 90, 95, 99, max]  312.756µs, 1.473ms, 1.439ms, 1.706ms, 1.9ms, 2.877ms, 6.977ms
Bytes In      [total, mean]                     0, 0.00
Bytes Out     [total, mean]                     0, 0.00
Success       [ratio]                           100.00%
Status Codes  [code:count]                      200:184  210:194  220:217  230:207  240:224  250:181  260:192  270:215  280:195  290:191  
Error Set:
```