@szuecs 2025-12-18
This change created a fix but tests that start with no endpoints and create multiple new endpoints were flaky because the tests did not change to make sure that we do not fadein but have an even distribution.

see also https://github.com/zalando/skipper/pull/3786/files#diff-ca886c0604353f1f9ef28c5d0dd9bc41ed172442c5dfe26c17c3aed440a180d5R493-R507